+++ MiaLogic/Manager/AanvraagManager.cs
@@ -97,495 +97,16 @@ namespace MiaLogic.Manager
 
             return highestAanvraagId;
         }
-        public static List<Afdeling> GetAfdelingen()
-        {
-            List<Afdeling> afdelingen = new List<Afdeling>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM Afdeling ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Afdeling afdeling = new Afdeling
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            afdelingen.Add(afdeling);
-                        }
-                    }
-                }
-            }
-
-            return afdelingen;
-        }
-        public static List<Dienst> GetDiensten()
-        {
-            List<Dienst> diensten = new List<Dienst>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM Dienst ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Dienst dienst = new Dienst
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            diensten.Add(dienst);
-                        }
-                    }
-                }
-            }
-
-            return diensten;
-        }
-        public static List<Prioriteit> GetPrioriteiten()
-        {
-            List<Prioriteit> prioriteiten = new List<Prioriteit>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM Prioriteit ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Prioriteit prioriteit = new Prioriteit
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            prioriteiten.Add(prioriteit);
-                        }
-                    }
-                }
-            }
-
-            return prioriteiten;
-        }
-        public static List<Financiering> GetFinancieringen()
-        {
-            List<Financiering> financieringen = new List<Financiering>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM FinancieringsType ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Financiering financiering = new Financiering
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            financieringen.Add(financiering);
-                        }
-                    }
-                }
-            }
-
-            return financieringen;
-        }
-        public static List<Investering> GetInvesteringen()
-        {
-            List<Investering> investeringen = new List<Investering>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM InvesteringsType ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Investering investering = new Investering
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            investeringen.Add(investering);
-                        }
-                    }
-                }
-            }
-
-            return investeringen;
-        }
-        public static List<string> GetFinancieringsjaren()
-        {
-            List<string> financieringsjaren = new List<string>();
-
-            // Adding financial years based on the current year
-            int currentYear = DateTime.Now.Year;
-            for (int i = 0; i < 5; i++)
-            {
-                string startingYear = (currentYear + i).ToString();
-                financieringsjaren.Add(startingYear);
-            }
-
-            return financieringsjaren;
-        }
-        public static List<Kostenplaats> GetKostenplaatsen()
-        {
-            List<Kostenplaats> kostenplaatsen = new List<Kostenplaats>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM Kostenplaats ORDER BY Naam ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            Kostenplaats kostenplaats = new Kostenplaats
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-
-                            kostenplaatsen.Add(kostenplaats);
-                        }
-                    }
-                }
-            }
-
-            return kostenplaatsen;
-        }
-        public static List<string> GetWieKooptHet()
-        {
-            List<string> aankoper = new List<string>();
-
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Voornaam + ' ' + Achternaam AS FullName FROM Aankoper ORDER BY FullName ASC";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        while (reader.Read())
-                        {
-                            string fullName = reader["FullName"].ToString();
-                            aankoper.Add(fullName);
-                        }
-                    }
-                }
-            }
-
-            return aankoper;
-        }
-
-        // Get...ById
-
-        public static Afdeling GetAfdelingById(int id)
-        {
-            Afdeling afdeling = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Naam FROM Afdeling WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                afdeling = new Afdeling
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Naam = reader["Naam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Dienst niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return afdeling;
-        }
-        public static Dienst GetDienstById(int id)
-    {
-        Dienst dienst = null;
-
-        try
-        {
-            using (SqlConnection connection = new SqlConnection(ConnectionString))
-            {
-                connection.Open();
-
-                string query = "SELECT Id, Naam FROM Dienst WHERE Id = @Id";
-
-                using (SqlCommand command = new SqlCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@Id", id);
-
-                    using (SqlDataReader reader = command.ExecuteReader())
-                    {
-                        if (reader.Read())
-                        {
-                            dienst = new Dienst
-                            {
-                                Id = Convert.ToInt32(reader["Id"]),
-                                Naam = reader["Naam"].ToString()
-                            };
-                        }
-                    }
-                }
-            }
-        }
-        catch (Exception)
-        {
-            Console.WriteLine("Het systeem kon de Dienst niet vinden, probeer het nog eens.");
-            throw;
-        }
-
-        return dienst;
-    }
-        public static Prioriteit GetPrioriteitById(int id)
-        {
-            Prioriteit prioriteit = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Naam FROM Prioriteit WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                prioriteit = new Prioriteit
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Naam = reader["Naam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Prioriteit niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return prioriteit;
-        }
-        public static Financiering GetFinancieringById(int id)
-        {
-            Financiering financieringsType = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Naam FROM FinancieringsType WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                financieringsType = new Financiering
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Naam = reader["Naam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Financiering niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return financieringsType;
-        }
-        public static Investering GetInvesteringById(int id)
-        {
-            Investering investeringsType = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Naam FROM InvesteringsType WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                investeringsType = new Investering
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Naam = reader["Naam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Investering niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return investeringsType;
-        }
-        public static Kostenplaats GetKostenplaatsById(int id)
-        {
-            Kostenplaats kostenplaats = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Naam FROM Kostenplaats WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                kostenplaats = new Kostenplaats
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Naam = reader["Naam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Kostenplaats niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return kostenplaats;
-        }
-        public static WieKooptHet GetAankoperById(int id)
-        {
-            WieKooptHet aankoper = null;
-
-            try
-            {
-                using (SqlConnection connection = new SqlConnection(ConnectionString))
-                {
-                    connection.Open();
-
-                    string query = "SELECT Id, Voornaam, Achternaam FROM Aankoper WHERE Id = @Id";
-
-                    using (SqlCommand command = new SqlCommand(query, connection))
-                    {
-                        command.Parameters.AddWithValue("@Id", id);
-
-                        using (SqlDataReader reader = command.ExecuteReader())
-                        {
-                            if (reader.Read())
-                            {
-                                aankoper = new WieKooptHet
-                                {
-                                    Id = Convert.ToInt32(reader["Id"]),
-                                    Voornaam = reader["Voornaam"].ToString(),
-                                    Achternaam = reader["Achternaam"].ToString()
-                                };
-                            }
-                        }
-                    }
-                }
-            }
-            catch (Exception)
-            {
-                Console.WriteLine("Het systeem kon de Aankoper niet vinden, probeer het nog eens.");
-                throw;
-            }
-
-            return aankoper;
-        }
+        
+       
+        
+        
+        
+        
+        
+
+            
+        
 
         // Aanvragen opslaan
         public static void SaveAanvraag(Aanvraag aanvraag, bool insert)
