+++ MiaLogic/Manager/AfdelingenManager.cs
@@ -1,0 +1,83 @@
+﻿using MiaLogic.Object;
+using System;
+using System.Collections.Generic;
+using System.Data.SqlClient;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MiaLogic.Manager
+{
+    public class AfdelingenManager
+    {
+        public static string ConnectionString;
+        public static List<Afdeling> GetAfdelingen()
+        {
+            List<Afdeling> afdelingen = new List<Afdeling>();
+
+            using (SqlConnection connection = new SqlConnection(ConnectionString))
+            {
+                connection.Open();
+
+                string query = "SELECT Id, Naam FROM Afdeling ORDER BY Naam ASC";
+
+                using (SqlCommand command = new SqlCommand(query, connection))
+                {
+                    using (SqlDataReader reader = command.ExecuteReader())
+                    {
+                        while (reader.Read())
+                        {
+                            Afdeling afdeling = new Afdeling
+                            {
+                                Id = Convert.ToInt32(reader["Id"]),
+                                Naam = reader["Naam"].ToString()
+                            };
+
+                            afdelingen.Add(afdeling);
+                        }
+                    }
+                }
+            }
+
+            return afdelingen;
+        }
+        public static Afdeling GetAfdelingById(int id)
+        {
+            Afdeling afdeling = null;
+
+            try
+            {
+                using (SqlConnection connection = new SqlConnection(ConnectionString))
+                {
+                    connection.Open();
+
+                    string query = "SELECT Id, Naam FROM Afdeling WHERE Id = @Id";
+
+                    using (SqlCommand command = new SqlCommand(query, connection))
+                    {
+                        command.Parameters.AddWithValue("@Id", id);
+
+                        using (SqlDataReader reader = command.ExecuteReader())
+                        {
+                            if (reader.Read())
+                            {
+                                afdeling = new Afdeling
+                                {
+                                    Id = Convert.ToInt32(reader["Id"]),
+                                    Naam = reader["Naam"].ToString()
+                                };
+                            }
+                        }
+                    }
+                }
+            }
+            catch (Exception)
+            {
+                Console.WriteLine("Het systeem kon de Dienst niet vinden, probeer het nog eens.");
+                throw;
+            }
+
+            return afdeling;
+        }
+    }
+}
