+++ MiaLogic/Manager/FinancieringenManager.cs
@@ -1,0 +1,83 @@
+﻿using MiaLogic.Object;
+using System;
+using System.Collections.Generic;
+using System.Data.SqlClient;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MiaLogic.Manager
+{
+    public class FinancieringenManager
+    {
+        public static string ConnectionString { get; set; }
+        public static List<Financiering> GetFinancieringen()
+        {
+            List<Financiering> financieringen = new List<Financiering>();
+
+            using (SqlConnection connection = new SqlConnection(ConnectionString))
+            {
+                connection.Open();
+
+                string query = "SELECT Id, Naam FROM FinancieringsType ORDER BY Naam ASC";
+
+                using (SqlCommand command = new SqlCommand(query, connection))
+                {
+                    using (SqlDataReader reader = command.ExecuteReader())
+                    {
+                        while (reader.Read())
+                        {
+                            Financiering financiering = new Financiering
+                            {
+                                Id = Convert.ToInt32(reader["Id"]),
+                                Naam = reader["Naam"].ToString()
+                            };
+
+                            financieringen.Add(financiering);
+                        }
+                    }
+                }
+            }
+
+            return financieringen;
+        }
+        public static Financiering GetFinancieringById(int id)
+        {
+            Financiering financieringsType = null;
+
+            try
+            {
+                using (SqlConnection connection = new SqlConnection(ConnectionString))
+                {
+                    connection.Open();
+
+                    string query = "SELECT Id, Naam FROM FinancieringsType WHERE Id = @Id";
+
+                    using (SqlCommand command = new SqlCommand(query, connection))
+                    {
+                        command.Parameters.AddWithValue("@Id", id);
+
+                        using (SqlDataReader reader = command.ExecuteReader())
+                        {
+                            if (reader.Read())
+                            {
+                                financieringsType = new Financiering
+                                {
+                                    Id = Convert.ToInt32(reader["Id"]),
+                                    Naam = reader["Naam"].ToString()
+                                };
+                            }
+                        }
+                    }
+                }
+            }
+            catch (Exception)
+            {
+                Console.WriteLine("Het systeem kon de Financiering niet vinden, probeer het nog eens.");
+                throw;
+            }
+
+            return financieringsType;
+        }
+    }
+}
