+++ MiaLogic/MiaLogic.csproj
@@ -42,21 +42,36 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Manager\AanvraagManager.cs" />
+    <Compile Include="Manager\AfdelingenManager.cs" />
+    <Compile Include="Manager\DienstenManager.cs" />
+    <Compile Include="Manager\FinancieringenManager.cs" />
+    <Compile Include="Manager\FinancieringsjaarManager.cs" />
     <Compile Include="Manager\FotoManager.cs" />
     <Compile Include="Manager\GebruikerManager.cs" />
     <Compile Include="Manager\GebruiksLogManager.cs" />
+    <Compile Include="Manager\InvesteringenManager.cs" />
+    <Compile Include="Manager\KostenplaatsManager.cs" />
     <Compile Include="Manager\LinkManager.cs" />
     <Compile Include="Manager\OfferteManager.cs" />
     <Compile Include="Manager\ParameterManager.cs" />
+    <Compile Include="Manager\PrioriteitManager.cs" />
     <Compile Include="Manager\RolManager.cs" />
+    <Compile Include="Manager\WieKooptHetManager.cs" />
     <Compile Include="Object\Aanvraag.cs" />
+    <Compile Include="Object\Afdeling.cs" />
+    <Compile Include="Object\Dienst.cs" />
+    <Compile Include="Object\Financiering.cs" />
     <Compile Include="Object\Foto.cs" />
     <Compile Include="Object\Gebruiker.cs" />
     <Compile Include="Object\GebruiksLog.cs" />
+    <Compile Include="Object\Investering.cs" />
+    <Compile Include="Object\Kostenplaats.cs" />
     <Compile Include="Object\Link.cs" />
     <Compile Include="Object\Offerte.cs" />
     <Compile Include="Object\Parameter.cs" />
+    <Compile Include="Object\Prioriteit.cs" />
     <Compile Include="Object\Rol.cs" />
+    <Compile Include="Object\WieKooptHet.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup />
