+++ MiaLogic/Object/Afdeling.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MiaLogic.Object
+{
+    public class Afdeling
+    {
+            public int Id { get; set; }
+            public string Naam { get; set; }
+    }
+}
