+++ MiaLogic/Object/WieKooptHet.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MiaLogic.Object
+{
+    public class WieKooptHet
+    {
+        public int Id { get; set; }
+        public string Voornaam { get; set; }
+        public string Achternaam { get; set; }
+    }
+}
